requirements.txt
Django==3.2.25
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.1
drf-yasg==1.21.7
psycopg2-binary==2.9.7
Pillow==10.0.0
requests==2.31.0
whitenoise==6.5.0
